created working generator
WARNING this generator has a try catch block command that will catch a Stack Overflow error and force the program to continue. This could freeze up your computer but I really wanted to see if it would work and it does took a long time but it did work out. There is clearly a lot of room for optimization and I think now that I have established the worst possible generator, it will only get better from here.