added the puzzle method
This will produce a puzzle by removing numbers and then filling them in 50 times and seeing if each attempt gives the same grid. Basically it checks for uniqueness in a puzzle and removes random elements. The only major run time complexity is the hidden complexity of the generator function as the puzzle method calls that multiple times. So optimization of generator.java is completely reliant on the generator method.