found problem with generator
The reason why the generator isn't efficient is because the grid contains method doesn't work properly. If I can fix that, the efficiency will improve greatly. Also, I optimized it a lot by not having it recursively call upon the generator method until it gets a correct sudoku puzzle. That took up a lot of unnecessary memory and always caused a stack overflow error