substituted file inputs with generator method
I got rid of the readPuzzle method and now Sudoku runs on the generator. I also added a feature where if you press h the program will give you a hint by giving you the number of the slot you are currently on. If you give up, the program will terminate and print out the solution of the puzzle. In the menu, I regulated difficulty levels by having if you choose easy, you get a puzzle with 40 numbers, medium gives 30, and hard, 20 numbers.