accounted for unsolvable sudoku puzzles
Two commits ago, I might have jumped the gun in saying that my generator has O(n^5). I incorrectly assumed that as long as you follow the rules of the sudoku puzzle, you will be able to fill it without any revisions. I then tested this out for myself and realized that it was very easy to create a puzzle that leads to an unsolvable contradiction. So, I edited my code such that if none of the numbers 1-9 work on a specific square, you set the block back to its original input and create a new arrangement of the sequence to put in. Then, if that doesn't work 5 times, start the grid over from the beginning. If it still is polynomial time, it would have an astronomically high coefficient but since I now have it start over from the beginning, it's sort of acts like a recursive algorithm as it starts from where the function first began giving it a very fast growing run time complexity. Thankfully, the generator isn't meant to be used in general cases so we don't really need to worry too much about efficiency.