got rid of generator put back readPuzzle
I wasn't making good enough progress with the generator so I put back the readPuzzle method. The final product for this project will probably be an interface but I'm also thinking of using Java to posts the time taken to complete the puzzle on Twitter. That way people can see how they fared against others and it can be a fun component of the project.