Final Project Commit
I added a couple of features to my project. You first have to add your name into the terminal for as an argument the main method will take in. Then, you can choose your difficulty: easy, medium, and hard and it will generate a random sudoku puzzle with that difficulty. This will bring up a sudoku puzzle where you can solve and be timed. Once you solve it, press escape and the program will tweet your name and the time taken to finish the puzzle to a twitter account I created called PsuedoSudoKhan. It will also print on the terminal the solution to the sudoku puzzle.
 To compile use this javac -cp lanterna.jar:twitter4j-core-4.0.7.jar: Sudoku.java and 
to run: java -cp lanterna.jar:twitter4j-core-4.0.7.jar: Sudoku